Changes in report generation
