+++ src/main/java/page/BasePage.java
@@ -1,16 +1,21 @@
 package page;
 
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
+import org.apache.commons.io.FileUtils;
+import org.openqa.selenium.*;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
+import java.io.File;
+import java.io.IOException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
 import java.util.List;
 
 public class BasePage {
 
     public static WebDriver driver;
+    private static final String fileSeparator = File.separator;
+    private static final String rootDirectory = System.getProperty("user.dir");
 
     public BasePage(WebDriver driver) {
         this.driver = driver;
@@ -44,5 +49,21 @@ public class BasePage {
     public void sendKeys(By by, String text) {
         waitUntilElementVisible(by);
         driver.findElement(by).sendKeys(text);
+    }
+
+    public static String takeScreenshot(String screenshotName) {
+        String timestamp = new SimpleDateFormat("yyyy-MM-dd HH-mm-ss").format(new Date());
+        TakesScreenshot ts = (TakesScreenshot) driver;
+        File source = ts.getScreenshotAs(OutputType.FILE);
+        String destination = rootDirectory + fileSeparator + "reports" + fileSeparator +
+                "html-report" + fileSeparator + "screenshots" + fileSeparator +
+                screenshotName + " - " + timestamp + ".png";
+        File finalDestination = new File(destination);
+        try {
+            FileUtils.copyFile(source, finalDestination);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return destination;
     }
 }
